WIP: Allow section to face right
Summary
-------
The section can now open to the right so slices are on left side of
circle.

Why
---
Need to be able to get slice positions for an arc open to the right.

This change addresses the need by
---------------------------------
Allow direction to be passed in when initializing dicer and
conditionally get angles and positions on other side of circle if
direction is "right".

Side effects
------------
Extra conditionals for each angle and position and negative total angle